:truck: Rename(fetch-users): 유저 데이터 가져오기 공통 컴포넌트로 변경
